refactor(core)!: decompose hypercache.go into domain modules and modernize ConcurrentMap iterator
Split the monolithic hypercache.go (~950 lines removed) into focused files:
- hypercache_construct.go: backend resolution and base construction
- hypercache_io.go: Set/Get/GetOrSet/GetMultiple/Remove/Clear/List
- hypercache_eviction.go: eviction configuration, loop, and triggers
- hypercache_expiration.go: expiration routine and trigger logic
- hypercache_dist.go: distributed metrics helpers (DistMetrics, ClusterOwners, etc.)

Replace ConcurrentMap's channel-based IterBuffered() with iter.Seq2-based
All(), eliminating fan-in goroutines and per-shard channel allocations.
Update all callers in dist_memory, dist_http_server, and inmemory backends.

Upgrade hash function from inlined FNV-1a to xxhash64 (XOR-folded to 32
bits) for better avalanche and ~1-3% speedup on longer keys.

Add bench-step3.txt with benchmark results on Apple M4 Pro.

BREAKING CHANGE: IterBuffered() and Tuple type removed from ConcurrentMap;
use All() iter.Seq2[string, *Item] instead.
